requirements.txt
Flask
pyzbar
Pillow
cryptography
web3
pymongo
PyMuPDF
pdf2image
qrcode
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
pytesseract
flask-bootstrap
fitz
frontend
requests